Refactored backend structure and switched to MongoDB
