+++ database/db.js
@@ -72,19 +72,40 @@
 
 // database/db.js
 
+// const mongoose = require("mongoose");
+// const { DATABASE_URL } = require("../config");
+
+// class DB {
+//   static async connect() {
+//     try {
+//       await mongoose.connect(DATABASE_URL, {
+//         useNewUrlParser: true,
+//         useUnifiedTopology: true,
+//       });
+//       console.log("✅ Connected to MongoDB");
+//     } catch (error) {
+//       console.error("MongoDB connection error:", error);
+//       process.exit(1);
+//     }
+//   }
+// }
+
+// module.exports = DB;
+
 const mongoose = require("mongoose");
-const { DATABASE_URL } = require("../config");
+const { MONGO_URI, MONGO_DB_NAME } = require("../config");
 
 class DB {
   static async connect() {
     try {
-      await mongoose.connect(DATABASE_URL, {
+      const fullUri = `${MONGO_URI}/${MONGO_DB_NAME}`;
+      await mongoose.connect(fullUri, {
         useNewUrlParser: true,
         useUnifiedTopology: true,
       });
       console.log("✅ Connected to MongoDB");
     } catch (error) {
-      console.error("MongoDB connection error:", error);
+      console.error("MongoDB connection error:", error.message);
       process.exit(1);
     }
   }
