+++ routes/dashboardRoutes.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+
+// Sample route
+router.get("/", (req, res) => {
+  res.send("User route working!");
+});
+
+module.exports = router;
